refactor: 혜택과 관련된 클래스 묶음
