+++ src/main/java/christmas/domain/config/benefit/BenefitConfig.java
@@ -1,4 +1,6 @@
-package christmas.domain.config;
+package christmas.domain.config.benefit;
+
+import christmas.domain.config.Menu;
 
 public class BenefitConfig {
     public static final Menu GIVEAWAY_ITEM = Menu.CHAMPAGNE;
